// src/app/shared/app/models/Claims/iclaims-filter.ts
import { FormControl } from "@angular/forms";
export interface IClaimsFilter {
  topStatusCount?: FormControl<number | null>;
  clientId?: FormControl<string | null>;
  clientName?: FormControl<string | null>;
  memberDriverName?: FormControl<string | null>;
  branch?: FormControl<string | null>;
  savedUser?: FormControl<string | null>;
  insurCompany?: FormControl<string | null>;
  claimType?: FormControl<string[] | null>;
  status?: FormControl<string[] | null>;
  subStatus?: FormControl<string[] | null>;
  claimNo?: FormControl<string | null>;
  policyNo?: FormControl<string | null>;
  insurCompClaimNo?: FormControl<string | null>;
  medID?: FormControl<string | null>;
  carNo?: FormControl<string | null>;
  chassisNumber?: FormControl<string | null>;
  nameofInjured?: FormControl<string | null>;
  lossLocation?: FormControl<string | null>;
  shipmentName?: FormControl<string | null>;
  insuredClaimNo?: FormControl<string | null>;
  blawbNo?: FormControl<string | null>;
  policyCertificateNo?: FormControl<string | null>;
  declarationNo?: FormControl<string | null>;
  paymentDetails?: FormControl<string | null>;
  wipNo?: FormControl<string | null>;
  invoiceNo?: FormControl<string | null>;
  dtpLossFrom?: FormControl<string | null>;
  dtpLossTo?: FormControl<string | null>;
  dtpCreatedOnFrom?: FormControl<string | null>;
  dtpCreatedOnTo?: FormControl<string | null>;
  paidAmount1?: FormControl<string | null>;
  paidAmount2?: FormControl<string | null>;
  underProcesingAmount1?: FormControl<string | null>;
  underProcesingAmount2?: FormControl<string | null>;
  combOperatorAmount?: FormControl<string | null>;
  combOperatorUnderProcessingAmount?: FormControl<string | null>;
}